Create LICENSE.md
As per https://community.smartthings.com/t/withdrawn-log-events-to-google-sheets/36719/115:
Charles_Schwer:
I have no plans at this time to come back. If you have downloaded the code, feel free to continue to use it. I released it with the Apache license v2 so there are no restrictions on continued use.

Thanks